refactor: wip: switching to `@zip.js/zip.js` (#18)
* export types used or returned by FileCollection
* re-export `configure` from `zip.js` as `zipJsConfigure`

String as binary is ambiguous, it could be:  

* basic base64
* dataUrl base64
* hexadecimal
* utf8 raw codepoints (could corrupt binary with conversion)

So we choose to drop string input support for `fromIum`. 

BREAKING CHANGE: remove `fromIum` input `string` and `NodeJS.ReadableStream` support.
If you used string you have to decode it to `Uint8Array`.
If you used Node.js Readable Streams, convert them to Web Stream with `Readable.toWeb`

fix: `.stream()` method on `sources` added by `FileCollection.appendSource`

It returned a promise rejecting an error because it used `.stream()` on a response (this method does not exists).  
Now it instantiates a stream, and the body response is piped to the stream

fix: file paths from `toIum` zip should not include `//` anymore